[HOPS-1506] HDFS-8218. Move classes that used by ClientProtocol into hdfs-client. Contributed by Haohui Mai.
