+++ hadoop-hdfs-project/hadoop-hdfs-client/src/main/java/org/apache/hadoop/hdfs/protocol/NSQuotaExceededException.java
@@ -21,46 +21,40 @@ package org.apache.hadoop.hdfs.protocol;
 import org.apache.hadoop.classification.InterfaceAudience;
 import org.apache.hadoop.classification.InterfaceStability;
 
-import java.io.IOException;
-
-/**
- * This exception is thrown when modification to HDFS results in violation
- * of a directory quota. A directory quota might be namespace quota (limit
- * on number of files and directories) or a diskspace quota (limit on space
- * taken by all the file under the directory tree). <br> <br>
- * <p/>
- * The message for the exception specifies the directory where the quota
- * was violated and actual quotas. Specific message is generated in the
- * corresponding Exception class:
- * DSQuotaExceededException or
- * NSQuotaExceededException
- */
 @InterfaceAudience.Private
 @InterfaceStability.Evolving
-public class QuotaExceededException extends IOException {
+public final class NSQuotaExceededException extends QuotaExceededException {
   protected static final long serialVersionUID = 1L;
-  protected String pathName = null;
-  protected long quota; // quota
-  protected long count; // actual value
-  
-  protected QuotaExceededException() {
-  }
 
-  protected QuotaExceededException(String msg) {
+  private String prefix;
+
+  public NSQuotaExceededException() {}
+
+  public NSQuotaExceededException(String msg) {
     super(msg);
   }
-  
-  protected QuotaExceededException(long quota, long count) {
-    this.quota = quota;
-    this.count = count;
+
+  public NSQuotaExceededException(long quota, long count) {
+    super(quota, count);
   }
-  
-  public void setPathName(String path) {
-    this.pathName = path;
-  }
-  
+
   @Override
   public String getMessage() {
-    return super.getMessage();
+    String msg = super.getMessage();
+    if (msg == null) {
+      msg = "The NameSpace quota (directories and files)" +
+      (pathName==null?"":(" of directory " + pathName)) +
+          " is exceeded: quota=" + quota + " file count=" + count;
+
+      if (prefix != null) {
+        msg = prefix + ": " + msg;
+      }
+    }
+    return msg;
+  }
+
+  /** Set a prefix for the error message. */
+  public void setMessagePrefix(final String prefix) {
+    this.prefix = prefix;
   }
 }
